+++ collector-exoplayer-v1/src/main/java/com/bitmovin/analytics/exoplayer/ExoPlayerAdapter.java
@@ -24,6 +24,7 @@ import com.bitmovin.analytics.enums.DRMType;
 import com.bitmovin.analytics.enums.PlayerType;
 import com.bitmovin.analytics.enums.VideoStartFailedReason;
 import com.bitmovin.analytics.error.ExceptionMapper;
+import com.bitmovin.analytics.exoplayer.manipulators.BitrateEventDataManipulator;
 import com.bitmovin.analytics.features.Feature;
 import com.bitmovin.analytics.license.FeatureConfigContainer;
 import com.bitmovin.analytics.stateMachines.PlayerState;
@@ -48,7 +49,6 @@ import com.google.android.exoplayer2.source.dash.manifest.DashManifest;
 import com.google.android.exoplayer2.source.hls.HlsManifest;
 import com.google.android.exoplayer2.source.hls.playlist.HlsMasterPlaylist;
 import com.google.android.exoplayer2.source.hls.playlist.HlsMediaPlaylist;
-import com.google.android.exoplayer2.trackselection.TrackSelection;
 import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -76,8 +76,8 @@ public class ExoPlayerAdapter
     private ExceptionMapper<Throwable> exceptionMapper = new ExoPlayerExceptionMapper();
     private final DeviceInformationProvider deviceInformationProvider;
     private DownloadSpeedMeter meter = new DownloadSpeedMeter();
+    private BitrateEventDataManipulator bitrateEventDataManipulator;
     private boolean isVideoAttemptedPlay = false;
-    private long previousQualityChangeBitrate = 0;
     private boolean isPlaying = false;
     private boolean isPaused = false;
 
@@ -95,6 +95,7 @@ public class ExoPlayerAdapter
         this.exoplayer.addListener(this);
         this.config = config;
         this.deviceInformationProvider = deviceInformationProvider;
+        this.bitrateEventDataManipulator = new BitrateEventDataManipulator(exoplayer);
         attachAnalyticsListener();
     }
 
@@ -122,6 +123,7 @@ public class ExoPlayerAdapter
     }
 
     private void startup(long position) {
+        bitrateEventDataManipulator.setFormatsFromPlayer();
         stateMachine.transitionState(PlayerState.STARTUP, position);
         isVideoAttemptedPlay = true;
     }
@@ -223,28 +225,6 @@ public class ExoPlayerAdapter
             }
         }
 
-        // Info on current tracks that are playing
-        if (exoplayer.getCurrentTrackSelections() != null) {
-            for (int i = 0; i < exoplayer.getCurrentTrackSelections().length; i++) {
-                TrackSelection trackSelection = exoplayer.getCurrentTrackSelections().get(i);
-                if (trackSelection != null) {
-                    Format format = trackSelection.getSelectedFormat();
-                    switch (exoplayer.getRendererType(i)) {
-                        case TRACK_TYPE_AUDIO:
-                            data.setAudioBitrate(format.sampleRate);
-                            break;
-                        case TRACK_TYPE_VIDEO:
-                            data.setVideoBitrate(format.bitrate);
-                            data.setVideoPlaybackHeight(format.height);
-                            data.setVideoPlaybackWidth(format.width);
-                            break;
-                        default:
-                            break;
-                    }
-                }
-            }
-        }
-
         data.setDownloadSpeedInfo(meter.getInfo());
 
         // DRM Information
@@ -263,17 +243,20 @@ public class ExoPlayerAdapter
             simpleExoPlayer.removeAnalyticsListener(this);
         }
         meter.reset();
+        bitrateEventDataManipulator.reset();
         stateMachine.resetStateMachine();
     }
 
     @Override
     public void resetSourceRelatedState() {
+        bitrateEventDataManipulator.reset();
         // no Playlist transition event in older version of collector (v1)
     }
 
     @Override
     public void registerEventDataManipulators(EventDataManipulatorPipeline pipeline) {
         pipeline.registerEventDataManipulator(this);
+        pipeline.registerEventDataManipulator(bitrateEventDataManipulator);
     }
 
     @Override
@@ -638,30 +621,50 @@ public class ExoPlayerAdapter
             long initializationDurationMs) {}
 
     @Override
-    public void onDecoderInputFormatChanged(EventTime eventTime, int trackType, Format format) {
+    public void onDecoderInputFormatChanged(
+            @NotNull EventTime eventTime, int trackType, @NotNull Format format) {
         try {
-            if ((this.stateMachine.getCurrentState() == PlayerState.PLAYING)
-                    || (this.stateMachine.getCurrentState() == PlayerState.PAUSE)) {
-                Log.d(
-                        TAG,
-                        String.format(
-                                "onDecoderInputFormatChanged: Bitrate: %d Resolution: %d x %d",
-                                format.bitrate, format.width, format.height));
-                if (format.bitrate == this.previousQualityChangeBitrate) {
-                    Log.d(TAG, "onDecoderInputFormatChanged: Skipping sample sending");
-                    return;
-                }
-                this.previousQualityChangeBitrate = format.bitrate;
-                if (this.stateMachine.isQualityChangeEventEnabled()) {
-                    long videoTime = getPosition();
-                    PlayerState originalState = this.stateMachine.getCurrentState();
-                    this.stateMachine.transitionState(PlayerState.QUALITYCHANGE, videoTime);
-                    this.stateMachine.transitionState(originalState, videoTime);
-                }
+            switch (trackType) {
+                case TRACK_TYPE_AUDIO:
+                    handleAudioInputFormatChanged(format);
+                    break;
+                case TRACK_TYPE_VIDEO:
+                    handleVideoInputFormatChanged(format);
+                    break;
             }
         } catch (Exception e) {
             Log.d(TAG, e.getMessage(), e);
         }
+    }
+
+    private void handleAudioInputFormatChanged(Format format) {
+        Log.d(TAG, String.format("onAudioInputFormatChanged: Bitrate: %d", format.bitrate));
+        long videoTime = getPosition();
+        PlayerState originalState = stateMachine.getCurrentState();
+        try {
+            if (stateMachine.getCurrentState() != PlayerState.PLAYING) return;
+            if (!stateMachine.isQualityChangeEventEnabled()) return;
+            if (!bitrateEventDataManipulator.hasAudioFormatChanged(format)) return;
+            stateMachine.transitionState(PlayerState.QUALITYCHANGE, videoTime);
+        } finally {
+            bitrateEventDataManipulator.setCurrentAudioFormat(format);
+        }
+        stateMachine.transitionState(originalState, videoTime);
+    }
+
+    private void handleVideoInputFormatChanged(Format format) {
+        Log.d(TAG, String.format("onVideoInputFormatChanged: Bitrate: %d", format.bitrate));
+        long videoTime = getPosition();
+        PlayerState originalState = stateMachine.getCurrentState();
+        try {
+            if (stateMachine.getCurrentState() != PlayerState.PLAYING) return;
+            if (!stateMachine.isQualityChangeEventEnabled()) return;
+            if (!bitrateEventDataManipulator.hasVideoFormatChanged(format)) return;
+            stateMachine.transitionState(PlayerState.QUALITYCHANGE, videoTime);
+        } finally {
+            bitrateEventDataManipulator.setCurrentVideoFormat(format);
+        }
+        stateMachine.transitionState(originalState, videoTime);
     }
 
     @Override