+++ CHANGELOG.md
@@ -7,10 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development v2
 
+### Fixed
+- ExoPlayer: wrong field to track audioBitrate
+- ExoPlayer: no quality changes if the quality did not change
+
 ### Added
 - Error detail tracking feature
 
 ## Development v1
+
+### Fixed
+- ExoPlayer: wrong field to track audioBitrate
+- ExoPlayer: no quality changes if the quality did not change
 
 ### Added
 - Error detail tracking feature